Ref: extracting routing walker.
